separate components, working keyboard
keyboard didn't highlight chars before, it is now fixed
